On some platforms HW ticks may overflow in uint32_t type.

+++ rtos/kernel/kernel.h
@@ -41,8 +41,8 @@
 #endif /* CMAKE_BUILD */
 #define MS_TO_TICK(a)               ((uint32_t)(((uint32_t)(a) * SOFT_TICKS_PER_SEC) / (1000)))
 #define TICK_TO_MS(a)               ((uint32_t)(((uint32_t)(a) * 1000) / (SOFT_TICKS_PER_SEC)))
-#define US_TO_HW_TICK(a)            ((uint32_t)(((uint32_t)(a) * HW_TICKS_PER_SEC) / (1000000)))
-#define HW_TICK_TO_US(a)            ((uint32_t)(((uint32_t)(a) * 1000000) / (HW_TICKS_PER_SEC)))
+#define US_TO_HW_TICK(a)            ((uint64_t)(((uint64_t)(a) * HW_TICKS_PER_SEC) / (1000000)))
+#define HW_TICK_TO_US(a)            ((uint64_t)(((uint64_t)(a) * 1000000) / (HW_TICKS_PER_SEC)))
 
 /* Some useful macros. */
 #define OFFSETOF(type, field)       ((int) &(((type *) 0)->field))
